+++ CHANGELOG.md
@@ -7,6 +7,8 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
+- Added unmatched airports to Markdown file
+
 ## [1.2.0] - 2025-07-25
 
 - Added NPM package main script for accessing data